Generate test cases more flexibly
This makes it possible to generate test cases either from a problem directory,
or from the root of the repository. E.g. both of the following now work:

    cd xgo
    go run leap/example_gen.go

and

    cd xgo/leap
    go run example_gen.go

This clarifies the assumption about where the x-common repository
is, making it relative to the gen/gen.go file rather than relative to
the present working directory.

The test_cases.go file is generated to the directory containing the
example_gen.go file.