Update large number test to increase security
Nothing in the [README] seems to indicate that numbers larger than 32
wouldn't be in the set of potential handshakes.

[README]: https://github.com/exercism/x-common/blob/master/secret-handshake.md

Practically speaking, a sophisticated attacker could brute-force so few
combinations (31 codes each with 8 possible handshakes = 248), so we'd
want to ensure that larger numbers with seemingly no relation to the
significant binary bits would also be possible codes.