Temporarily ignore the error-handling exercise
The error-handling problem includes a test fixture that is binary data, and the API is not handling
it correctly, causing the JSON parser to blow up when attempting to construct the payload.

I started addressing the problem with a workaround in the API, but realized that this would leak
complexity into parts of the API unrelated to just fetching exercises (restore, list, skip, etc),
and decided that it was better to not do that.

Solving this correctly is going to impact the CLI in some way, since we will need to perform some base64
decoding before writing file contents, and we'll need to think about how to do this in a backwards-
compatible way. Perhaps we can make it so that a broken exercise gets a placeholder README that says
that they need to upgrade to get that problem.

I need to think about it more.