react: Show line numbers in assertCellValue
Having a common function has a problem in that `go test` will report the
failed line number as the line number of the common function, not of the
calling test. For example, all failures would be reported at line 30.

Since it is valuable to show the user what line caused the test failure,
we'll use the runtime package to inspect the caller of assertCellValue
and have assertCellValue report that.