react: Add more callback tests
In particular we add a test for not calling a callback if dependencies
change but the computed value does not change, as discussed in #189.

We also add a test of multiple callbacks being added and removed, since
that was not tested before. Confirmed that @soniakeys solution crashes
with this additional test, requested in #189.

Finally, a test for remove idempotence.

Test version is bumped for this.