ledger: Make initial implementation pass test
The readme for the ledger exercise claims that the current code
consistently passes the tests. Unfortunately, on my machine it
consistently does not pass the tests. Every time I try, I get this:

```
--- FAIL: TestFormatLedgerSuccess (0.00s)
        ledger_test.go:267: FormatLedger for input named "credit and
debit" was expected to return...
                Date_______|_Description_______________|_Change
                01/01/2015_|_Buy_present_______________|______($10.00)
                01/02/2015_|_Get_present_______________|_______$10.00_

                ...but returned...
                Date_______|_Description_______________|_Change
                01/02/2015_|_Get_present_______________|_______$10.00_
                01/01/2015_|_Buy_present_______________|______($10.00)
```

Looking into the initial implementation, the implementation is creating
a goroutine for each line of the ledger, and the main goroutine is
collecting the formatted lines from a channel that all the goroutines
are sending to. Unfortunately, there are no guarantees as to the order
in which the goroutines will send on the channel. In fact on my machine
they consistently send the results in the wrong order, so I always fail
the test.

Well, I can fix that, we'll just make each goroutine aware of the index
of the line it's formatting, and the main goroutine needs to reorder
them in the right order. Now it will consistently pass the tests.

The alternative, of course, would be to modify the readme to say that
the ledger might not even pass the tests all the time.