ledger: remove multibyte characters
The discussion we had in #195 indicates that Unicode introduces a whole
host of issues that are better suited for their own exercise. Let's keep
the ledger exercise focused on refactoring. This avoids a potential
source of confusion of "why is it cutting at 24 instead of 22?"

ledger test version is bumped for this.