triangle: Ensure no pair of triangle types are equal
As reported in #207, the triangle tests can be solved by a trivial empty
implementation that simply sets all four triangle types equal to one
another. We should prevent false positives, and we can do so by checking
that Equ != Iso, etc.

I verified that 1) the trivial solution fails the new test, and 2) the
example solution passes the test.

Fixes #207