triangle: Allow triangles where a+b == c
These are degenerate triangles since they're really just lines, but
they're still good enough under the triangle inequality (which
explicitly allows a+b == c) so let's go ahead and test for them.

Test for isosceles and scalene added. Equilateral is impossible.

Unfortunately, now we are at risk for failing a test case of (3, Inf,
Inf), since 3 + Inf == Inf, and Inf < Inf is false whereas previously
the test was Inf <= Inf which is true. So we might falsely declare this
triangle as isosceles when it should be NaT. So we need to add another
test to make sure that our largest side isn't +Inf.

Closes #231