triangle: un-nest equilateral check
Since they're sorted, if a == c, then this is an equilateral triangle.
If not, then we only need to check whether a == b or b == c, which will
make it isosceles, else it's scalene.

This flattens the switch/case structure, and compresses two isosceles
cases into one, so overall it seems easier to understand.