paasio: Document required concurrency guarantees
As discussed in #143, since the tests are going so far as to ensure that
the counters are still accurate in the face of concurrent operations, it
is wise to document these guarantees as well. This also clarifies that
there need be no guarantees on interleaving of Read/Write calls, another
discussion point in #143.