paasio: Use channels instead of mutexes
The discussion in #143 suggests that we might like to avoid mutexes
unless we have a good reason for why channels can't do the job. This
commit implements a channel-based approach to counters. A single
goroutine is responsible for updating the bytes/ops counters. All other
goroutines send requests to it on the write channel.

Since an unprotected read of the counter variables would be a data race
(which is rightfully caught by Go's race detector!), reads must be
protected as well. This is done by sending channels into a channel, at
which point the values are sent on the channel.

Unanswered question: Now when we no longer need a ReadCounter or a
WriteCounter, we have no way to get rid of its corresponding goroutine.
Ideally they'd have Close functions, but we currently do not ask for
this, and there are complications if we do: the type system doesn't let
us embed ReadCounter+WriteCounter if they both have Close... we'd have
to duplicate the function definitions which is very unfortunate.