travis: move stub files instead of building `example`
Some exercises have stub files to ease newcomers to Go. Some other
exercises have initial solutions that need to be refactored. In both
cases, these initial files cannot be compiled when we are running Travis
tests, as they would cause conflicting definitions. We avoid this
problem by deleting the stub files.

The alternative was to use build constraints to exclude building those
files.

This approach has the advantage that students trying to write code and
understand the stub files won't have to puzzle through what these build
constraints actually mean. The disadvantage is that this assumes a
naming convention for the stub files and now track maintainers are bound
to this convention.

As a side effect, two exercises that explicitly specify a constraint had
to have these constraints removed since Travis no longer specifies it.

This arose out of discussion in #248.