poker: deduplicate cards from hands
I'm still undecided on whether duplicate cards in the same hand should
be an error, but it's at least reasonable to me that tests should only
test one thing at a time. So regardless of whether they are an error, we
can have a single test that tests that behavior instead of a bunch of
them.

Fixes #250