hamming: Check that err is an error
Tested by first making the example `Distance` return an `*int` instead
of an `error` (without changing the test).

The test passed, indicating that the additional line in the test is
necessary.

Then, the line was added to the test, which caused a compilation error
as hoped. So this is correctly checking that the `err` return value is
indeed an `error`. This is the standard technique in Go for asserting
that a value implements an interface.

This arises out of discussion in #275. This will be an example that will
be used when providing guidance on how to enact the changes for other
exercises.