remove unnecessary stub files
As part of discussion in #279, we discovered that these exercises have
three-line stub files that aren't particularly useful (just a package
name and test version).

If we suppose that the purpose of stub files is to help out newcomers to
Go, and that we gradually ease them off of stub files, it seems less
useful to have stub files for exercises that come late in the track. By
that time, students should be able to figure out package names and test
versions themselves.

Affects:
circular-buffer
crypto-square
custom-set
food-chain
paasio
tournament

Closes #279