test-without-stubs: Use trap to cleanup while keeping exit status
We'd like to ensure the cleanup happens no matter how the script is
exited.

More importantly, this moves commands so that `go test` is once again
the last command, so that the exit status of the script is the exit
status of that command.

This is important because otherwise Travis will falsely report test
success when instead the test was failing.

This script was introduced in fa897488609899b977adf95c693f3535c6315321
as part of #249 an initiative to not have build constraints which were
confusing to students.

Luckily we have not had any test failures since then, but it would be
good to restore the right behavior.