series: rename Frist to UnsafeFirst
As noted in #329 and #175, it's confusing to call the function Frist
because it looks like an unintentional typo.

As I understand it, it was intentional and meant to show that this
version of the function (that doesn't signal errors with a second return
value) is unsafe in some cases.

The real version should be called First (its rightful name), so we had
to name the initial version something different.

So I propose UnsafeFirst rather than Frist which makes it clear that:

* The function is unsafe.
* The name is intentional.

Perhaps it will also inspire more questions of "since it's unsafe, how
can I make it safe?" which seems a better motivation for doing the bonus
than "let's fix a typo".

Closes #329.