house: re-implement exercise, fixes #242
Remove `Embed()` function, and enable a non-recursive solution. The test
file still suggests using recursion as a method for solving this
problem, and both recursive and non-recursive solutions are documented
in the `example.go` file. Reading suggestions for recursion in Go have
also been added.

An extra error check has also been added to the Song() test, to check
for the number of verses in the solution. This is to avoid unhelpful test
failure messages when a whole verse is missing, such as:

```
house_test.go:156: Song() line 77 = "", want ""
```

now reads:

```
house_test.go:139: Song() has 11 verses, want 12 verses
```