SETUP.md: move to exercises/TRACK_HINTS.md
As https://github.com/exercism/exercism.io/issues/3240 states, we would
like to rename SETUP.md.

As we can see from #417, the unfortunate name of SETUP.md has caused
confusion on what is supposed to go inside it.

https://github.com/exercism/trackler/pull/17 has added the necessary
support for TRACK_HINTS.md, so we can now move this file.