exercises: remove executable bits from *.go files
The generator was writing these files as executable.

One will find that an attempt to execute them directly will not succeed
(they have no shebang line, so a Unix shell might naively attempt to
execute them as shell script, with hilariously bad results)

The generator has been fixed to not write them as executable, so we'll
just fix up the existing files that are executable.