nucleotide-count: Decrease benchmark test runtime
Improve the running time for 'go test -bench .'
by removing StartTimer/StopTimer references.

StopTimer and StartTimer are for excluding startup costs
from a benchmark, but for this benchmark there is no
startup overhead logic to exclude, and the running time
is severely increased by the timer operations.

The observed improvement for the example
is roughly 9x from about 14 seconds to about 1.5 seconds.