travis: No longer test 1.5.4 and 1.4.3
I made the original decision to test additional versions in #316, on the
knowledge that travis only runs five jobs at a time, assuming that the
restriction is only across our own repo.

However, as https://github.com/exercism/x-common/issues/539 seems to
imply, the restriction is across the repos of the entire organisation.

It is thus probably better citizenship to test only what we need, and we
clearly did not need 1.5 and 1.4 since we would not really have cared if
they failed.