gen.go: collect canonical-data.json remotely
If the gen.go file cannot find a local copy of the required
canonical-data.json file it will then attempt to collect the json data
from the raw canonical-data file in the x-common repository on GitHub.

This enables the person who is regenerating the exercise's cases_test.go
file to not have to have an up to date version of the x-common repo on
their local machine. There is always the potential this local copy will
not be up to date with current changes.
However, as @petertseng has pointed out: when "working on a new generator
and I want to run it very often to see whether it works. I wouldn't want
to have to continually fetch the new JSON file from the network every
time I try that, since it will lengthen the feedback loop and that makes
development less pleasant", so the gen.go file will continue to first use
a local repo if one is available.

see https://github.com/exercism/xgo/pull/612#discussion_r111094977