pascals-triangle: add benchmarks (#649)
* pascals-triangle: add benchmarks

Added two benchmarks to the pascals triangle test.

One notable side-effect is that I moved the size of the test from hard-coded to
an independent variable so that it may be used as the size for the fixed length
test and upper bound for the incremental size test. Also renamed the test cases
from the hardcoded size name.