series: Update package name to better reflect exercise name
This excercise had a package name which was odd compared to the other
exercises in the track: it did not contain any part of the exercise name.
in fact it seems to be the name of the core concept behind the probable
solution in Go.

This also updates the test version since the test would now be in a
different package.