gen/gen.go: improve error support
Update gen/gen.go package to give additional information
when an error occurs.

This change outputs an additional log output line to indicate
the context of the error, either during template Execute
or during the source Format step.

Also, when the error occurs during the source Format,
go ahead and output the generated source to the cases_test.go
file anyway with an extra line appended at the end which contains the
error message from the source Format.

Being able to examine the erroneous source text makes it easier
to understand the specific issue with the template/generator output.

Show either [SUCCESS] or [ERROR] output: /path/to/cases_test.go
in a log message after saving the generated source.