README: clarify intentions of Example tests
Issue #420 asked us to consider using Example tests for some exercises.
This was discussed in #420 & #675 and resulted in implementations in #568
This commit clarifies the view of the current maintainers that example
tests are not necessarily useful as a replacement for data-driven tests,
but rather as a way to illustrate and verify the API of an exercise.

closes #420