README: link to contributing docs in docs repo
Because documentation items are being moved to the docs repo and thus
being removed from x-common's CONTRIBUTING file, it makes less sense to
link to the latter.

The section about the contents of the contributing guide no longer was
accurate, so most of it was simply removed and replaced with "ways to
get involved".