gen: Add support for `property` annotation
This allows us to specify types for each key depending on the `property`
key of a JSON object rather than having to use the least common
denominator type (which is often `interface{}`).

This change is compatible with generators that do not use this
annotation (and JSON files that have only one property will not need
it).

Discussions on this topic:
https://github.com/exercism/go/pull/674#discussion_r117160215
https://github.com/exercism/discussions/issues/155#issuecomment-306352911