clock 1.0.1 (#706)
* clock: Support arbitrary add test names

In clock 1.0.1 the tests with property `add` were split up into two
groups. As a result, we can't call both of them `addTests`.
https://github.com/exercism/problem-specifications/pull/740

The solution we will decide on for now is to base the variable name on
the first word of the description. Since they are "Add minutes" and
"Subtract minutes" they will become addTests and subtractTests.

This does not yet update to clock 1.0.1, simply paves the way for doing
so. This change is compatible with both 1.0.0 and 1.0.1, however it will
result in a group named `testTests` if running with 1.0.0.

It's acknowledge that we are now kind of dependent on the descriptions,
but this was better than hard-coding the group names in, perhaps. We can
deal with it later on if the descriptions change to something like "the
following tests test adding" which would case our name to be `theTests`
which is a terrible name.

* clock 1.0.1: Split add tests into add and subtract

https://github.com/exercism/problem-specifications/pull/740