Add .editorconfig file
The [.editorconfig file](http://editorconfig.org/)  is compatible with most development environments and is used
to setup project-based consistent text editing style. For this repo,
gofmt already handles this for the .go files. The .editorconfig file
is for consistency in the the [config.json](config.json) file. This would solve
the tabs/spaces problem (see #714) automatically for anyone with an editorconfig
aware environment.

Closes #722