bank-account: more info about expected solution + mutex links (#710)
Added a little more detail on the expected solution, concurrency, and not
allowing the balance to go below 0 on withdrawals.

Looking at the submitted solutions some users were not discovering Mutexes
on their own. Thought it may be helpful to add links to the godocs/tutorials
concerning Mutexes if people have no idea where to start on the solution.