error-handling: Add lengthy description of expected solution (#707)
This is sort of a tricky problem if you are new to Go based on where it
currently falls in the exercises list. Seemed like people are getting lost in
the "what" of the problem instead of focusing on the "how". The README is
necessarily generic because error handling differs greatly between
languages. This updates the solution description in the tests file with more
detail of the expected behaviour of the Use function in the solution. It also
provides links to the language concepts which will be used in the solution and
some comments in the common file.

Resolves #703