Update introductory copy
This adds the 'blurb' to the config file, and tweaks the about doc. I really like the sentence
in the about document that talks about how gophers like to describe the language. It feels very friendly
and colloquial, and makes a great hook to pull people in to read more.

So I put that first, and then tweaked the second sentence simply for flow, so that it fit after the opening
paragraph.