Restructure configfile for Nextercism, difficulties 1-2
Our initial stab at the reconfiguration for Nextercism incorporating
elements of #794 and #798. The "comment" field is an experimental
addition to the configuration to make remembering each exercise a
little easier while working on the file.

Analysis of the first two levels of difficulty led to some
re-assigning of those values. Exercises with stubs moved to
difficulty. Exercises without a stub moved to 2.