hello-world: Simplify to be less overwhelming
I recently realized how incredibly overwhelming the massive
wall of text is in the hello world exercise.

This simplifies it down to the barest of bare bones, which I think
feels friendlier and easier to complete.

In addition to removing comments, this also removes the testVersion
constant and the example test.

My reasoning for removing the testVersion is primarily that most people
don't give feedback on the Hello, World exercise, and if they do, then
they probably have seen it enough to know that it got simplified.

Secondarily, when we launch the new v2 prototype publically, we will
be able to get rid of the testVersions entirely, since we will
always show the mentors the exact test suite that the user downloaded.

I don't anticipate needing to change hello-world between now and then,
so we shouldn't see any confusion amongst the few people who may give
feedback on hello-world.

Also, I'm going to implement two-fer (the exercise that replaced the
conditional version of hello-world), and give it an example test.