Change some exercises to be 'bonus' exercises
Bonus exercises are those that are not part of the main *core* exercise
structure, they are available as soon as the hello-world exercise is
submitted.

See https://github.com/exercism/go/pull/808#discussion_r134311568

Other exercises have also been moved in response to the above.

The bonus exercises; `bob`, `leap` & `raindrops`, have all been edited
so they are more suitable for someone who has just completed the
`hello-world` exercise. They all have simplified stubs, with the test
versioning (soon to be removed completely) moved and commented on to be
less important before the explanation in `hamming`.