maintainers: add my bio as an alumnus (#820)
As noted in https://github.com/exercism/go/issues/719 I cannot currently
fulfill my duties as a maintainer of this track, so it is right and
proper to mark myself as an alumnus as a result.

I will do as Katrina did in https://github.com/exercism/go/pull/797 and
submit a bio for myself. I think this is suitable for showing on the
website (by the way, Katrina's currently isn't being shown).