gen/gen.go: correct dirExercise determination
Using Go 1.9 with test case generators breaks
the placement of the output file, cases_test.go.
Specifically, on Linux it is output to ../cases_test.go which is
relative to your current directory.

The bug is related to the fact that with Go 1.9, this line

 	if _, path, _, ok := runtime.Caller(2); ok {

when called from init() within a package
has a different result than 1.8.3 and earlier.
Path is returned as "<autogenerated>" with 1.9,
and dirExercise ends up being set to "..".

The correct solution for both new and previous Go versions
is to utilize the same technique of runtime.Caller(),
but from a known call stack position apart from the init() function.

The change moves the logic to set dirExercise
from init() into the Gen() function in order to
set a correct value for the exercise directory regardless
of the current working directory when running the generator.

Also, add some commentary explanation of how dirExercise is determined.

Resolves #848