[nextercism] Restructure config.json (#835)
* [nextercism] Restructure config.json

This PR stems from #717, #794, #805, #813, #825

Closes #828, #829, #830

It should be noted that while my name is attached to this commit, this
PR would not have been possible without the hard work and insights of
Tom Leen (@tleen).

For Exercism v2, codename Nextercism, the structure of exercises will
change. It will no longer be a linear progression, rather there will be
a series of core exercises that will lead on from one another with
mentor approval. These core exercises will in turn unlock other
exercises that can be completed as and when. This change implements the
new exercise progression structure.