generator: handle multi-line commit - issue #874
When remote problem-specifications repository is used
as the source of the canonical-data.json, and the commit entry
retrieved has a multi-line message, the generator
uses the entire multi-line message in the source comment output;
this causes a compile error in the generated cases_test.go.

Correct bug by using only the first line of the commit message
in returned string from getRemoteCommit.

Closes #874