simple-linked-list: improve benchmark runtime
Running go test -bench on this exercise is taking
far too much time (several minutes).

Improve the benchmark routines by
doing the list creation outside the loop
and using ResetTimer; this applies to routines
    BenchmarkNewList,
    BenchmarkListSize,
    BenchmarkListToArray, and
    BenchmarkListReverse.

For BenchmarkListPush and BenchmarkListPop,
use an inner loop to measure 100 instances
of the respective Push or Pop operations.
This approach reduces the overhead runtime in relation
to the operations being measured.

Resolves #1073