Added tests and example implementation for concat
Concat is hard to do without type polymorphism because []IntSlice
is not the same type as [][]IntSlice so while a concat function
that operates over type []A' where A' could be IntSlice or []IntSlice
would be straightforward, in go it would either have to be implemented
as separate methods or over []interface{} with a function that
dispatches on a type assertion. At least as far as I can determine,
being a go newbie. :)