fix: Update test cases for outdated exercises.
1. Why is this change neccesary?
Because some exercises like isogram allowed you to use a `break` a statement
inside a for loop and having all the tests pass, this can lead to some confusion
between `break` and `continue` statements that are inside a for loop.

With the current test suite one could do something like this:

```
for i, letter := range word {
		letter = unicode.ToLower(letter)
		if !unicode.IsLetter(letter) {
			break
		}
		if strings.ContainsRune(word[i+1:], letter) {
			return false
		}
	}
```

And have all the test pass.

This could be problematic as the `break` statement gives a better benchmark but
it's because of a coincidence with the test cases.

Other exercises had similar problems.

2. How does it address the issue?
The test cases for all of the exercises were updated.

3. What side effects does this change have?
We're up-to-date with the latest test cases! :)