Use %q to make white space differences visible
Some student complained that the tests failed but he couldn't see any
difference between "expected" and "actual".
After looking into the issue (by changing the %s to %q in this test), I
see that an extra "\n" was appended to the result, which is hard to tell
from the output.