Fixing Test / Main Packaging (#1519)
Test seems to expect the 'system under test' to be in a numbers package, but it's in a `cars` package, same as the test -- when working on this locally, the simplest thing seemed to be to remove the import and drop the qualified method names, which was in keeping with the way `savings-account` was written. There are other paths, but this is simple.