zero_zilch_nada: remove ambiguous reference to empty (#1500)
* zero_zilch_nada: remove ambiguous reference to empty

In Go, there can be a difference between a type's zero value and its
empty value. For example, consider the following code snippet:

```go
func main() {
	var s1 []string
	fmt.Printf("%#v\n", s1)
	if s1 == nil {
		fmt.Println("s1 is equal to nil")
	}

	s2 := []string{}
	fmt.Printf("%#v\n", s2)
	if s2 == nil {
		fmt.Println("s2 is equal to nil")
	}
}
```

The variable `s1` will be set to its zero value (`nil`) whereas the
variable `s2` will be set to a slice of strings with zero elements. This
behavior makes the term "empty" ambiguous and may confuse students as they
continue learning about Go.

In addition to code updates, the exercise text was updated to introduce
the `var` keyword as well as various verbiage updates.

* zero_zilch_nada: update related concept documentation

Updated the documentation for the nil and zero-values concepts to better
reflect the new `zero_zilch_nada` exercise content. Long term, I'd
probably recommend merging the two concepts into one but for now keeping
their content the same ensures that students will learn the relevant
content regardless of which conept they click on.

* added missing period character

* Update concept documentation and exercise code

Updated the concept documentation to focus on zero values.

Changed the exercise to be a bit more involved than the previous
one-liner.

* Add missing period

* Update nil concept

* Add census exercise

* Implement pull request feedback

* Use inline linking

* Implement pull request feedback

* Remove nil concept

* Update zero-values concept content

* Update census exercise

- Ensure the links in the hints are inline.
- Ensure the introduction includes information on struct types.
- Restore the exercise code to an unfinished state.
- Update the testing code.

* Use relative links instead of absolute links

* Update prerequisites

* Ensure tests run in the same package as exercise