Unify examples between the exercise and the concept.
Also tidy up the shared text to be formatted the same for easier comparison/future changes.

Change if/else if/else to append to a string to make early return no longer a better way to write the code.
This is option (c) from #1699 which is my preference, but I'm happy to change it if others disagree.