Extended Instructions for AddItem Function
Based on the test suite, the expectation from the AddItem function is to add the input item to the bill and if its already present, to increase its quantity by unit amount. The current instructions mention only the former part of the requirement and leave out the latter. Included an additional line to fix this.