Fix broken link in nucleotide-count's description
The link at the end of the description goes to a GitHub page that is out
of date. In addition, I think it makes more sense to link to the leap
exercise on Exercism's own site rather than to a raw markdown document
on GitHub.

I've updated the link accordingly.