Add missing test cases for Annalyn's Infiltration
In this exercise the instructions say that the prisoner MUST be
awake in order to be rescued, as if they are sleeping they'll be
startled by the rescue attempt, which will alert the guards.

Today we saw a solution that would have allowed Annalyn to rescue
a sleeping prisoner if she didn't have her dog along.

This adds a couple of extra cases for that scenario.