lasagna: Fix ElapsedTime doc comment
Instead of the total time needed to prepare + bake a lasagna, which only
depends on the amount of layers, ElapsedTime calculates the time that
has passed until a specific moment in the baking process. This matches
the function signature, as it depends on the amount of layers and the
minutes it has been in the oven so far.